refactor: 주문 생성, 주문 트래킹 클래스 분리
